requirements.txt
pyzotero
PyMuPDF
watchdog
pyyaml
requests
anthropic
